# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Telecom Management - API Documentation",
        "version": "1.0"
    },
    "paths": {
...
        "/api/cucmcdrs/search/{column}/{search}": {
            "get": {
                "tags": [
                    "CDR - CUCM History"
                ],
                "summary": "List CDR Records by Date Range",
                "description": "",
                "operationId": "searchCDR",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "column",
                        "in": "path",
                        "description": "Column to Search",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "search",
                        "in": "path",
                        "description": "Search",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/api/cucmcdrs/calls_with_loss_by_daterange": {
            "post": {
                "tags": [
                    "CDR - CUCM History"
                ],
                "summary": "List CDR Records by Date Range",
                "description": "",
                "operationId": "list_calls_by_date_range",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "start",
                        "in": "formData",
                        "description": "Example: 5/10/2017",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "end",
                        "in": "formData",
                        "description": "Example: 5/10/2017",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/api/cucmcdrs/list_last_24hr_calls_by_number_search": {
            "post": {
                "tags": [
                    "CDR - CUCM History"
                ],
                "summary": "List Last 24hr CDR Records by Number",
                "description": "",
                "operationId": "list_calls_by_date_range",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "search",
                        "in": "formData",
                        "description": "Number to Search by",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        },
...
        "/api/activitylogs/bydates": {
            "post": {
                "tags": [
                    "Logging and Permissions"
                ],
                "summary": "Get Logs from Dates",
                "description": "",
                "operationId": "list_calls_by_date_range",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "start",
                        "in": "formData",
                        "description": "Example: 5/10/2017",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "end",
                        "in": "formData",
                        "description": "Example: 5/10/2017",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        },
...
        "/api/sonus5kcdrs/search/{column}/{search}": {
            "get": {
                "tags": [
                    "CDR - Sonus History"
                ],
                "summary": "List CDR Records by Date Range",
                "description": "",
                "operationId": "searchCDR",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "column",
                        "in": "path",
                        "description": "Column to Search",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "search",
                        "in": "path",
                        "description": "Search",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/api/sonus5kcdrs/callsbydaterange": {
            "post": {
                "tags": [
                    "CDR - Sonus History"
                ],
                "summary": "List CDR Records by Date Range",
                "description": "",
                "operationId": "list_calls_by_date_range",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "start",
                        "in": "formData",
                        "description": "Example: 5/10/2017",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "end",
                        "in": "formData",
                        "description": "Example: 5/10/2017",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/api/sonus5kcdrs/calls_with_loss_by_daterange": {
            "post": {
                "tags": [
                    "CDR - Sonus History"
                ],
                "summary": "List CDR Records by Date Range",
                "description": "",
                "operationId": "list_calls_by_date_range",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "start",
                        "in": "formData",
                        "description": "Example: 5/10/2017",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "end",
                        "in": "formData",
                        "description": "Example: 5/10/2017",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        },
...
        "/api/sonus/cdrs": {
            "get": {
                "tags": [
                    "Management - Sonus - Onbox CDRs"
                ],
                "summary": "Get CDRs",
                "description": "",
                "operationId": "backups",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorized user"
                    }
                }
            }
        },
...
    },
...
}